chore: baseline of prior work — Shopify cursor pagination/retry, inventory tracking, product search, AI platform endpoints, fulfillment push fixes

+++ nexus-oms-backend/src/main/java/com/nexus/oms/entity/NxProductMapping.java
@@ -21,10 +21,10 @@ public class NxProductMapping {
     private UUID tenantId;
 
     @Column(name = "bc_product_id", nullable = false)
-    private Integer bcProductId;
+    private Long bcProductId;
 
     @Column(name = "bc_variant_id")
-    private Integer bcVariantId;
+    private Long bcVariantId;
 
     @Column(name = "bc_sku")
     private String bcSku;
